Stat rebalance nightmare vortex

See merge request KartKrew/Kart!2053

+++ src/d_player.h
@@ -690,8 +690,8 @@ struct player_t
 
 	SINT8 drift;			// (-5 to 5) - Drifting Left or Right, plus a bigger counter = sharper turn
 	fixed_t driftcharge;	// Charge your drift so you can release a burst of speed
-	UINT8 driftboost;		// (0 to 125) - Boost you get from drifting
-	UINT8 strongdriftboost; // (0 to 125) - While active, boost from drifting gives a stronger speed increase
+	UINT16 driftboost;		// (0 to 125 baseline) - Boost you get from drifting
+	UINT16 strongdriftboost; // (0 to 125) - While active, boost from drifting gives a stronger speed increase
 
 	UINT16 gateBoost;		// Juicebox Manta Ring boosts
 	UINT8 gateSound;		// Sound effect combo
